Decompile func0f0bd358

+++ src/include/game/game_0b63b0.h
@@ -53,7 +53,7 @@ s16 func0f0bc478(void);
 u32 func0f0bc4c0(void);
 s16 func0f0bc554(void);
 s16 func0f0bc7ac(void);
-s32 func0f0bcaa0(void);
+s16 func0f0bcaa0(void);
 s16 func0f0bce24(void);
 f32 func0f0bd358(void);
 void func0f0bd3c4(void);
